+++ crack-core/src/handlers/mod.rs
@@ -4,6 +4,7 @@ pub mod idle;
 pub mod serenity;
 pub mod track_end;
 pub mod voice;
+pub mod voice_chat_stats;
 
 pub use self::event_log::handle_event;
 pub use self::idle::IdleHandler;
